Add files via upload
Refined evaluation function to include more positional and tactical information. Seems okay, but it doesn't yet account for threats on friendly pieces in the evaluation step. Expanded the Projection class to do most of the work of generating moves, in the future it will calculate influence on a square from either side using these and help maintain sets of attackers and defenders for each piece.